refactor: flatten elements source modules
Move the elements package glitch modules into the src root so the package structure matches its current scope.
