+++ packages/elements/src/index.ts
@@ -1,19 +1,19 @@
-import { GojibakeGlyphElement } from "./glitch/gojibake-glyph-element.js";
-import { GojibakeGlyphFragmentElement } from "./glitch/gojibake-glyph-fragment-element.js";
+import { GojibakeGlyphElement } from "./gojibake-glyph-element.js";
+import { GojibakeGlyphFragmentElement } from "./gojibake-glyph-fragment-element.js";
 
-export type { GojibakeGlyphLayout } from "./glitch/gojibake-glyph-element.js";
-export { GojibakeGlyphElement } from "./glitch/gojibake-glyph-element.js";
+export type { GojibakeGlyphLayout } from "./gojibake-glyph-element.js";
+export { GojibakeGlyphElement } from "./gojibake-glyph-element.js";
 export {
   DUAL_FRAGMENT_REGIONS,
   type FragmentRegion,
   GojibakeGlyphFragmentElement,
   type PlacementMode,
   QUAD_FRAGMENT_REGIONS,
-} from "./glitch/gojibake-glyph-fragment-element.js";
-export { GlitchRenderer } from "./glitch/renderer.js";
-export type { Config, DisplayState } from "./glitch/state-factory.js";
-export { GlitchStateFactory } from "./glitch/state-factory.js";
-export { pickGlyphForChar } from "./glitch/utils.js";
+} from "./gojibake-glyph-fragment-element.js";
+export { GlitchRenderer } from "./renderer.js";
+export type { Config, DisplayState } from "./state-factory.js";
+export { GlitchStateFactory } from "./state-factory.js";
+export { pickGlyphForChar } from "./utils.js";
 
 /** カスタム要素を document に登録する。重複登録は避ける。 */
 export function registerGojibakeElements(): void {
